backend/requirements.txt
# backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI/LLM
langchain==0.3.26
langchain-openai==0.3.24
langchain-community==0.3.26

# Vector Database
langchain-pinecone==0.2.8

# Document Processing
pdfplumber==0.10.3
sentence-transformers==4.1.0 
pypdf==5.6.1

# Utilities
python-dotenv==1.1.0
pydantic>=2.7.4,<3.0.0

# Additional Libraries
sentence-transformers==4.1.0 

# Note: CORS is built into FastAPI, no separate package needed